+++ src/app/app.component.ts
@@ -1,14 +1,18 @@
-import {Component} from '@angular/core';
+import {AfterViewInit, Component, ElementRef, ViewChild} from '@angular/core';
 
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.scss']
 })
-export class AppComponent {
-  title = 'shop';
-  clear = false;
-  handleEvent = (event) => {
-    this.clear = event;
+export class AppComponent implements AfterViewInit {
+  @ViewChild('appTitle') appTitle: ElementRef
+  name = 'Basketball Shop';
+
+  ngAfterViewInit() {
+    console.log(this.appTitle.nativeElement)
+    this.appTitle.nativeElement.textContent = this.name
   }
+
+
 }