+++ src/app/modules/products/components/product/product.component.ts
@@ -1,0 +1,16 @@
+import {Component, EventEmitter, Input, Output,} from '@angular/core';
+import {ProductModel} from "../../models/product.model";
+
+@Component({
+  selector: 'app-product',
+  templateUrl: './product.component.html',
+  styleUrls: ['./product.component.scss']
+})
+export class ProductComponent {
+  @Input() product: ProductModel
+  @Output() addProduct: EventEmitter<ProductModel> = new EventEmitter<ProductModel>()
+
+  onClick(product){
+    this.addProduct.emit(product)
+  }
+}