+++ src/app/modules/products/models/product.model.ts
@@ -1,6 +1,7 @@
-export interface Product {
+export interface ProductModel {
   name: string;
   price: number;
   isAvailable: boolean;
   categories: string;
+  id: number
 }