+++ changelog.md
@@ -1,1 +1,7 @@
-Task 1. Introduction 
+###Task 1. Introduction 
+
+###Task 2. Components
+1. add modules
+2. add services
+3. add directive
+4. implements  @Input(), @Output(),  @ViewChild, @HostBinding, @HostListener